requirements.txt
annotated-types==0.7.0
asttokens==3.0.1
attrs==25.4.0
black==25.11.0
blinker==1.9.0
cachelib==0.13.0
cattrs==25.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorcet==3.1.0
comm==0.2.3
dash==3.2.0
dash-ag-grid==32.3.2
dash-bootstrap-components==2.0.4
dash-bootstrap-templates==2.1.0
dash-extensions==2.0.4
dash-leaflet==1.1.3
dash-svg==0.0.12
dataclass-wizard==0.35.1
debugpy==1.8.17
decorator==5.2.1
dotenv==0.9.9
editorconfig==0.17.1
executing==2.2.1
flask==3.1.2
flask-caching==2.3.1
geobuf==2.0.0
html5lib==1.1
idna==3.11
importlib-metadata==8.7.0
ipykernel==7.1.0
ipython==9.7.0
ipython-pygments-lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
jsbeautifier==1.15.4
jupyter-client==8.6.3
jupyter-core==5.9.1
lxml==6.0.2
markupsafe==3.0.3
matplotlib-inline==0.2.1
more-itertools==10.8.0
mypy-extensions==1.1.0
narwhals==2.7.0
nest-asyncio==1.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
patsy==1.0.1
pip==25.3
platformdirs==4.5.0
plotly==6.3.1
plotly-express==0.4.1
prompt-toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
pure-eval==0.2.3
pyarrow==22.0.0
pydantic==2.12.4
pydantic-core==2.41.5
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
requests==2.32.5
requests-cache==1.2.1
retrying==1.4.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
statsmodels==0.14.5
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
wcwidth==0.2.14
webencodings==0.5.1
werkzeug==3.1.3
zipp==3.23.0
